Adding GraphQL and JWT Authentication
This addresses portions of issue #2

The GraphQL configuration has been attached to the express server,
and authentication logic has been added to it as middleware.
The user's decoded JWT is attached to this context and will only
be available if the user is authenticated.

The resolvers are currently empty, but the schema has been added
and the server is running.