Added controllers and resolvers for the following graphQl types:
This commit is a work in progress towards issue #3 and is working on
implementing the resolvers as defined in the typeDef file.

The following queries and mutations were added:
- Queries:
  - queryMe

- Mutations:
  - addUser
  - loginUser

Basic tests have been written for the userControllers and they are
currently all passing. Code coverage could use some improvement but,
most of it is repetitive code that is being tested and covered by
the tests that are currently in place, but in other areas. We know
mongoose works, we know the jwt works, we know the bcrypt works, and we
know we can create a user and query a user only if we are authenticated.