GraphQL Queries and Mutations are still a work in progress.
This commit is a work in progress towards issue #3.

The following mutations were added implemented:
  - deleteUser
  - editUser

Existing tests were updated a bit for the new mutations and we have
methods to:
   (C)reate a new user
   (R)ead a single user,
   (U)pdate a user
   (D)elete a user.

This essentially finishes off the UserModel proper. The next step is
to implement the barcode lookup feature and the products model. So we
can create lists and add things to it.