Add userProduct mutation has been created.
This commit is a work in progress towards issue #3 and completes the
addUserProduct mutation.

This mutation is used:

when a user adds a product to their list. The mutation takes in a
barcode and looks for an existing product and user product in the db.
If they don't exist, it will create them. If they do exist, it will
return the existing user product. This mutation isn't to be used
directly by the user. It was intended to be used as an API for adding
an item to their default or selected list from other methods.