Finished adding the remaining queries and mutations as defined in the
GraphQL schema. And ran tests in ApolloPlayground. This commit
essentially finishes issue #3. The Queries and Mutations that a client
needs to use have been defined in the Sandbox and tested. Most of
the controllers had tests written but all that testing needs to be
refactored so we don't have to repeat a lot of the same logic over
and over. Not all tests have been written, it was dragging down the
development process. But I will get to them.