Updated the pre-existing login mutation
This has been updated to accept a deviceKey
parameter to allow a user's TrashScan device
to post scanned items to their default lists.

A generateDeviceKey mutation has also been
created. The model includes a function to
hash this value before saving it and exposes
a verify method that can be used to determine
if a non-hashed deviceKey matches ours. Keys
are only good for 30 days and generating a
new key automatically invalidates the old one.